requirements.txt
python-chess
pygraphviz
graphviz
cairosvg
networkx
sh
requests
requests_cache
termcolor
ipython
